requirements.txt
astroid==2.6.5
asttokens==2.2.1
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.2
boto3==1.26.3
botocore==1.29.3
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle==2.2.0
configparser==5.3.0
cryptography==37.0.4
cycler==0.10.0
decorator==5.1.1
dnspython==2.2.1
docker==6.0.1
eli5==0.11.0
executing==1.2.0
flake8==3.9.2
fonttools==4.38.0
future==0.18.2
graphviz==0.17
hyperopt==0.2.5
idna==3.4
importlib-metadata==4.6.1
iniconfig==1.1.1
ipython==8.5.0
isort==5.9.2
jedi==0.18.2
Jinja2==3.0.1
jmespath==1.0.1
joblib==1.0.1
kaleido==0.2.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lightgbm==3.3.3
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mlxtend==0.19.0
mypy==0.910
mypy-extensions==0.4.3
networkx==2.8
nose==1.3.7
numpy==1.21.5
packaging==21.3
pandas==1.4.2
paramiko==2.11.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.9.0
pluggy==1.0.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py4j==0.10.9.5
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
Pygments==2.14.0
pylint==2.9.5
pymongo==4.3.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyspark==3.3.1
pytest==7.1.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
scar==4.3.0
scikit-learn==1.1.2
scipy==1.8.1
six==1.16.0
sklearn==0.0
stack-data==0.6.2
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
threadpoolctl==2.2.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
traitlets==5.8.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.12
wcwidth==0.2.5
websocket-client==1.4.2
wrapt==1.12.1
xgboost==1.6.1
zipp==3.5.0
